backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==3.0.0
PyJWT==2.8.0
pymongo>=4.6.3,<5.0
python-multipart==0.0.6
minio==7.1.11
opencv-python-headless==4.8.0.74
python-dotenv==1.0.0
stripe==7.10.0
faiss-cpu
numpy<2
Pillow
scikit-learn
sentence-transformers
bcrypt
ultralytics
torch
torchvision
facenet-pytorch
matplotlib
tqdm
scikit-image
imagehash
huggingface_hub

# Report Generation Module Dependencies
llama-cpp-python
weasyprint
reportlab
markdown
jinja2
